requirements.txt
Flask
python-dotenv
openai
weasyprint
jinja2